Added filter tests

+++ orders/views.py
@@ -15,9 +15,8 @@ def basket(request):
     data = request.GET
     user = request.user
     basket_items = Basket.objects.all()
-    if 'remove' in request.GET:
+    if 'remove' in request.GET and Basket.objects.filter(pk=data['remove']):
         Basket.objects.filter(pk=data['remove']).delete()
-        print('item removed')
     total = 0
     if 'id' in request.COOKIES:
         item = request.COOKIES['id']
@@ -79,7 +78,6 @@ def basket_add(request):
         else:
             customer_item = Customer.objects.get(user_hash=cookie_id,user=request.user)
     else:
-        print('no cookie')
         cookie_id = md5(('customer' + str(time()) + str(random.random())).encode('utf8')).hexdigest()
         resp.set_cookie('id', cookie_id)
         if request.user.is_anonymous():
@@ -92,18 +90,14 @@ def basket_add(request):
         data = request.GET
         item_dish = Dish.objects.get(pk=int(data['item']))
         dish_added = Basket.objects.filter(dish=item_dish, order=None, customer__user_hash=customer_item.user_hash).exists()
-        print(dish_added)
         if dish_added:
             item = Basket.objects.get(dish=item_dish, order=None, customer__user_hash=customer_item.user_hash)
             price_item = item_dish.price
             item.quantity += int(data['quantity'])
             item.price = price_item*item.quantity
             item.save()
-            print('basket increased')
-            print(item.quantity)
         else:
             total_price = item_dish.price*int(data['quantity'])
             basket_item = Basket(dish=item_dish, quantity=int(data['quantity']), price=total_price, customer=customer_item,)
             basket_item.save()
-            print('basket added')
     return resp
